Move message chain lookup into API
